+++ CHANGES.md
@@ -13,5 +13,7 @@
 
 - [ADD] Windows に対応する
   - @melpon
+- [ADD] `SoraClientBuilder::ice_server_url_configurer` を追加する
+  - @melpon
 
 ### misc